Rename LoginController to LoginRepository
