+++ app/src/main/java/com/crypto/katip/RegisterActivity.java
@@ -13,7 +13,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.crypto.katip.controllers.LoginController;
+import com.crypto.katip.login.LoginRepository;
 import com.crypto.katip.viewmodels.register.RegisterFormState;
 import com.crypto.katip.viewmodels.register.RegisterResult;
 import com.crypto.katip.viewmodels.register.RegisterViewModel;
@@ -57,7 +57,7 @@ public class RegisterActivity extends AppCompatActivity {
             @Override
             public void onChanged(RegisterResult registerResult) {
                 if (registerResult.getSuccess() != null) {
-                    LoginController.getInstance().login(registerResult.getSuccess(), getApplicationContext());
+                    LoginRepository.getInstance().login(registerResult.getSuccess(), getApplicationContext());
                     startActivity(new Intent(RegisterActivity.this, HomeActivity.class));
                     finish();
                 } else {
